+++ js/Demo/demoMod.js
@@ -11,8 +11,8 @@ let modInfo = {
 
 // Set your version in num and name
 let VERSION = {
-	num: "2.3.5",
-	name: "Cooler and Newer Edition",
+	num: "2.π",
+	name: "Incrementally Updated",
 }
 
 let changelog = `<h1>Changelog:</h1><br>