+++ js/utils/NumberFormating.js
@@ -6,7 +6,7 @@ function exponentialFormat(num, precision, mantissa = true) {
         m = new Decimal(1)
         e = e.add(1)
     }
-    e = (e.gte(10000) ? commaFormat(e, 0) : e.toStringWithDecimalPlaces(0))
+    e = (e.gte(1e9) ? format(e, 1) : (e.gte(10000) ? commaFormat(e, 0) : e.toStringWithDecimalPlaces(0)))
     if (mantissa)
         return m.toStringWithDecimalPlaces(precision) + "e" + e
     else return "e" + e