+++ js/utils/save.js
@@ -70,6 +70,8 @@ function getStartLayerData(layer) {
 		layerdata.best = new Decimal(0);
 	if (layerdata.resetTime === undefined)
 		layerdata.resetTime = 0;
+	if (layerdata.forceTooltip === undefined)
+		layerdata.forceTooltip = false;
 
 	layerdata.buyables = getStartBuyables(layer);
 	if (layerdata.clickables == undefined)