+++ changelog.md
@@ -1,8 +1,9 @@
 # The Modding Tree changelog:
 
 - Completely reworked tooltips. Shift-click a node to force its tooltip to stay displayed. (And hopefully finally fixed flickering!)
-- Added 2 new input components, text-input and slider.
+- Added text-input and slider components.
 - The red layer highlight will not appear before a layer is unlocked.
+- Node symbols can use HTML.
 
 # v2.π.1 - 4/7/21
 - Fixed formatting for some larger numbers.