+++ changelog.md
@@ -1,13 +1,16 @@
 # The Modding Tree changelog:
 
+## v2.4: Rationalized Edition
 - Completely reworked tooltips. Shift-click a node to force its tooltip to stay displayed. (And hopefully finally fixed flickering!)
 - Added text-input and slider components.
 - Added the ability to toggle respec confirmations.
 - Added custom respec confirmation messages.
 - The red layer highlight will not appear before a layer is unlocked.
 - Added unlocking hotkeys.
+- You no longer need to supply 'rows' and 'cols' for any Big Features.
 - Node symbols can use HTML.
 - Added documentation for the respec button.
+- The version number no longer contains special characters or irrational numbers.
 
 # v2.π.1 - 4/7/21
 - Fixed formatting for some larger numbers.